updated join page banner to link to application
