+++ global.css
@@ -5,7 +5,6 @@ main {
   width: 100vw;
   height: 100vh;
   overflow-x: hidden;
-
   background-size: 135vh;
 } 
 
@@ -36,13 +35,20 @@ p {
 }
 
 .index-main {
+  display:grid;
+  grid-template-columns: 100vw;
   grid-template-rows: 110vh 14vh;
 }
 .index-description {
   grid-template-columns: 48fr 52fr;
+  height: 75vh;
 }
 .index-description-text {
   margin-left: 2rem;
+}
+
+.description-container {
+  justify-content: center;
 }
 
 .about {
@@ -55,6 +61,35 @@ p {
 .join-content {
   grid-template-columns: 60vw;
   height: 20vh;
+}
+.zoom-button {
+  transition: transform .2s;
+}
+.zoom-button:hover {
+  transform: scale(1.1);
+  color: #ff9086;
+}
+.underline-button {
+  position: relative;
+}
+.underline-button:hover {
+  color: #ff9086;
+}
+.underline-button:before {
+  content: "";
+  position: absolute;
+  width: 100%;
+  height: 2px;
+  bottom: 0;
+  left: 0;
+  background-color: #ff9086;
+  visibility: hidden;
+  transform: scaleX(0);
+  transition: all 0.3s ease-in-out;
+}
+.underline-button:hover:before {
+  visibility: visible;
+  transform: scaleX(1);
 }
 .foot {
   height: 125px;
@@ -80,16 +115,21 @@ p {
 
   /* home/index page */
   .index-main {
-    grid-template-rows: 280vw 40vw;
+    grid-template-rows: 120vh auto;
   }
 
   .index-description {
     grid-template-columns: 100%;
+    height: 100vh;
   }
 
   .index-description-text {
     margin-left: 0;
     margin-top: 1.8rem;
+  }
+
+  .description-container {
+    justify-content: flex-start;
   }
 
   /* about page */
@@ -136,7 +176,7 @@ p {
   }
   /* join page layout */
   .join {
-    grid-template-rows: 300vw 100vw 40vw;
+    grid-template-rows: 110vh 50vh 40vh;
   }
   .join-content {
     grid-template-columns: 80vw;
